Add files via upload
This program compresses and decompresses text files according to the Huffman encoding algorithm. The algorithm performs a unique compression for each text it's given by assigning the least number of bits to represent the most frequent characters in the file (allowing you to represent some characters with as few as 2 or 3 bits as opposed to representing every character with 8). Coded in C++